backend/diagnosis/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
pydantic>=2.10.0
pydantic-settings>=2.7.0
google-cloud-aiplatform>=1.80.0
google-cloud-storage>=2.19.0
google-genai>=1.12.0
google-adk>=1.0.0
